Got the template working
It actually works now, and it only does one thing: turn a sql query into
a dataset profile with whylogs. In order to do any of the other things I
want to do I'll have to switch this over to a flex template which is a
bit more work. I'll just focus on making sure that people can actually
use this first.

Last blocker there is replacing the use of DatasetProfile with
ResultSet, and maybe also carrying the Dict<datetime, ResultSet> concept
over so its able to profile a handful of days at once.